// frontend/lib/traffic-commissioner/traffic-commissioner.type.ts
// import mongoose, { Document, Schema } from "mongoose";

// // communication types
// export enum CommunicationType {
//   Email = "Email",
//   PhoneCall = "Phone Call",
//   Letter = "Letter",
// }

// // Define and export an interface representing a TrafficCommissionerCommunication document
// export interface ITrafficCommissionerCommunication extends Document {
//   type: CommunicationType;
//   contactedPerson: string;
//   reason: string;
//   communicationDate: Date;
//   attachments?: mongoose.Types.ObjectId[];
//   comments?: Date;
//   standAloneId?: mongoose.Types.ObjectId;
//   createdBy?: mongoose.Types.ObjectId;
// }

// // Define the TrafficCommissionerCommunication schema
// const TrafficCommissionerCommunicationSchema: Schema<ITrafficCommissionerCommunication> =
//   new Schema(
//     {
//       type: {
//         type: String,
//         enum: Object.values(CommunicationType),
//         required: true,
//       } /* Type of communication */,
//       contactedPerson: {
//         type: String,
//         required: true,
//       },
//       reason: {
//         type: String,
//         required: true,
//       },
//       communicationDate: {
//         type: Date,
//         required: true,
//       },
//       attachments: [
//         {
//           type: Schema.Types.ObjectId,
//           ref: "Document", // Reference from Document model
//         },
//       ],
//       comments: {
//         type: Date,
//       },
//       standAloneId: {
//         type: Schema.Types.ObjectId,
//         ref: "User",
//       },
//       createdBy: {
//         type: Schema.Types.ObjectId,
//         ref: "User", // Reference from User model
//       },
//     },
//     { timestamps: true, versionKey: false },
//   );

// // Create the TrafficCommissionerCommunication model
// const TrafficCommissionerCommunication =
//   mongoose.model<ITrafficCommissionerCommunication>(
//     "TrafficCommissionerCommunication",
//     TrafficCommissionerCommunicationSchema,
//   );

// // Export the TrafficCommissionerCommunication model
// export default TrafficCommissionerCommunication;

export interface trafficCommissionerRow {
  _id: string;
  type: string;
  contactedPerson: string;
  reason: string;
  communicationDate: string;
  attachments?: string[];
  comments?: string;
  standAloneId?: string;
  createdBy?: string;
}

export interface trafficCommissionerListResponse {
  communications: trafficCommissionerRow[];
  totalData: number;
  totalPages: number;
}

export interface CreateTrafficCommissionerInput {
  type: string;
  contactedPerson: string;
  reason: string;
  communicationDate: string;
  attachments?: string[];
  comments?: string;
  standAloneId?: string;
}

export interface UpdateTrafficCommissionerInput {
  type?: string;
  contactedPerson?: string;
  reason?: string;
  communicationDate?: string;
  attachments?: string[];
  comments?: string;
}